refactor: CarName 관련 생성자 오버로딩 모두 삭제, 그에 맞게 테스트 코드 수정
