+++ src/main/java/racingcar/domain/CarName.java
@@ -1,9 +1,5 @@
 package racingcar.domain;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-
 public record CarName(String value) {
     private static final String NAME_DELIMITER = ",";
     private static final int NAME_MAX_LENGTH = 4;
@@ -16,12 +12,5 @@ public record CarName(String value) {
         if (nameValue.length() > NAME_MAX_LENGTH || nameValue.isEmpty()) {
             throw new IllegalArgumentException("[ERROR] 자동차 이름의 길이가 너무 깁니다.");
         }
-    }
-
-    public static List<CarName> parseFrom(String inputNames) {
-        return Arrays.stream(inputNames.split(NAME_DELIMITER))
-                .filter(name -> !name.isEmpty())
-                .map(CarName::new)
-                .collect(Collectors.toList());
     }
 }
