+++ src/main/java/racingcar/domain/Cars.java
@@ -2,7 +2,6 @@ package racingcar.domain;
 
 import java.util.*;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import racingcar.domain.numbergenerator.NumberGenerator;
 
 public class Cars {
@@ -12,10 +11,6 @@ public class Cars {
         cars = carNames.stream()
                 .map(name -> new Car(name, numberGenerator))
                 .collect(Collectors.toList());
-    }
-
-    public Cars(String inputNames, NumberGenerator numberGenerator) {
-        this(CarName.parseFrom(inputNames), numberGenerator);
     }
 
     public void allTryMove() {
