+++ src/test/java/racingcar/domain/CarTest.java
@@ -14,27 +14,34 @@ class CarTest {
     @Test
     void 이름_검증_테스트1() {
         String name = "david";
-        assertThatThrownBy(() -> new Car(name, new RandomNumberGenerator()))
+
+        assertThatThrownBy(() -> new Car(new CarName(name), new RandomNumberGenerator()))
                 .isInstanceOf(IllegalArgumentException.class);
     }
+
     @Test
     void 이름_검증_테스트2() {
         String name = "";
-        assertThatThrownBy(() -> new Car(name, new RandomNumberGenerator()))
+
+        assertThatThrownBy(() -> new Car(new CarName(name), new RandomNumberGenerator()))
                 .isInstanceOf(IllegalArgumentException.class);
     }
+
     @Test
     void 이름_입력_테스트() {
         String name = "a";
-        Car car = new Car(name, new RandomNumberGenerator());
+
+        Car car = new Car(new CarName(name), new RandomNumberGenerator());
+
         assertThat(car.getName()).isEqualTo(name);
     }
+
     @Test
     void 자동차_이동_성공_테스트() {
         String name1 = "dave";
         String name2 = "paul";
-        Car car1 = new Car(name1, successNumberGenerator);
-        Car car2 = new Car(name2, successNumberGenerator);
+        Car car1 = new Car(new CarName(name1), successNumberGenerator);
+        Car car2 = new Car(new CarName(name2), successNumberGenerator);
 
         car1.tryMove();
         car1.tryMove();
@@ -48,8 +55,8 @@ class CarTest {
     void 자동차_이동_실패_테스트() {
         String name1 = "dave";
         String name2 = "paul";
-        Car car1 = new Car(name1, failureNumberGenerator);
-        Car car2 = new Car(name2, failureNumberGenerator);
+        Car car1 = new Car(new CarName(name1), failureNumberGenerator);
+        Car car2 = new Car(new CarName(name2), failureNumberGenerator);
 
         car1.tryMove();
         car1.tryMove();
@@ -63,8 +70,8 @@ class CarTest {
     void 자동차_위치_동일_테스트() {
         String name1 = "dave";
         String name2 = "paul";
-        Car car1 = new Car(name1, successNumberGenerator);
-        Car car2 = new Car(name2, successNumberGenerator);
+        Car car1 = new Car(new CarName(name1), successNumberGenerator);
+        Car car2 = new Car(new CarName(name2), successNumberGenerator);
 
         car1.tryMove();
         car1.tryMove();
